+++ src/agent/onefuzz-agent/src/tasks/coverage/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-pub mod generic;
-pub mod libfuzzer_coverage;
-pub mod recorder;
-pub mod total;